Set up routing and template views
