+++ bandrclient/src/MainView.js
@@ -1,28 +1,8 @@
-import { useState, useEffect } from 'react';
-import logo from './logo.svg';
 import './App.css';
-import { Button } from 'reactstrap';
-import auth from './data/firebaseAuth';
-import { checkUserCreatedInDB, getAllUsers, getSingleUser, signInUser, signOutUser } from './data/userData';
-import UserCard from './components/UserCard';
+import Routing from './routes';
 
 function MainView() {
   // const [user, setUser] = useState(null);
-  const [userList, setUserList] = useState([]);
-  const [match, setMatch] = useState({});
-
-  useEffect(() => {
-  let isMounted = true;
-
-    getAllUsers().then((response) => {
-      if (isMounted){
-        setUserList(response)
-      }},);
-    return () => {
-      isMounted = false
-    }
-  }, [userList]);
-  
 
 //   useEffect(() => {
 //     auth.onAuthStateChanged(async (response) => {
@@ -51,23 +31,11 @@ function MainView() {
 //         }
 //     });
 // }, []);
-
-const findMatch = () => {
-let random = Math.floor(Math.random() * userList.length + 1);
-console.warn(random);
-getSingleUser(random).then(response => setMatch(response))
-};
   
   return (
     <div className="App">
       <header className="App-header">
-        <p>
-          Welcome to Bandr!
-        </p>
-        <UserCard user={match} />
-        <Button onClick={signInUser}>Sign In</Button>
-        <Button onClick={signOutUser}>Sign Out</Button>
-        <Button onClick={findMatch}>Find A Match</Button>
+      <Routing></Routing>
       </header>
     </div>
   );
