+++ bandrclient/src/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 import MainView from './MainView';
@@ -7,7 +8,9 @@ import MainView from './MainView';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
+    <Router>
     <MainView />
+    </Router>
   </React.StrictMode>
 );
