+++ bandrclient/src/views/DirectMessageView.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function DirectMessageView() {
+  return (
+    <div>
+        <h3>DirectMessageView</h3>
+    </div>
+  )
+}
