+++ bandrclient/src/views/EditView.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
-export default function UserView() {
+export default function EditView() {
   return (
     <div>
-      <h3>UserView</h3>
+        <h3>EditView</h3>
     </div>
   )
 }
