+++ bandrclient/src/views/MessagesView.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function MessagesView() {
+  return (
+    <div>
+        <h3> MessagesView</h3>
+    </div>
+  )
+}
