+++ bandrclient/src/views/SettingsView.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SettingsView() {
+  return (
+    <div>
+        <h3>SettingsView</h3>
+    </div>
+  )
+}
